# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register/warga": {
      "post": {
        "operationId": "AuthController_registerWarga",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWargaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register as Warga",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register/linmas": {
      "post": {
        "operationId": "AuthController_registerLinmas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterLinmasDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a Linmas Officer (Koordinator Only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login for all roles",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Refresh Access Token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout and clear FCM Token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/attendance/status": {
      "get": {
        "operationId": "AttendanceController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Today Attendance Status (Android Home)",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/geofence": {
      "get": {
        "operationId": "AttendanceController_getGeofence",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Master Geofence Configuration",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/clock-in": {
      "post": {
        "operationId": "AttendanceController_clockIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ClockInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clock-In Attendance with Photo Upload",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/clock-out": {
      "post": {
        "operationId": "AttendanceController_clockOut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ClockOutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clock-Out Attendance with Photo Upload",
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/recap": {
      "get": {
        "operationId": "AttendanceController_getRecap",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Monthly Attendance Recap",
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/incidents": {
      "post": {
        "operationId": "IncidentsController_createIncident",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Incident Report (Warga Only)",
        "tags": [
          "Incidents"
        ]
      },
      "get": {
        "operationId": "IncidentsController_getIncidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get All Incident List",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/active": {
      "get": {
        "operationId": "IncidentsController_getActiveIncidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Active Incidents Only",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/my": {
      "get": {
        "operationId": "IncidentsController_getMyIncidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Track My Incident Reports (Warga Only)",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{incident_id}": {
      "get": {
        "operationId": "IncidentsController_getIncidentDetail",
        "parameters": [
          {
            "name": "incident_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Incident Detail (Includes Rating & Private URL)",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{incident_id}/claim": {
      "post": {
        "operationId": "IncidentsController_claimIncident",
        "parameters": [
          {
            "name": "incident_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim Incident (\"Saya Tangani\")",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{incident_id}/status": {
      "patch": {
        "operationId": "IncidentsController_updateIncidentStatus",
        "parameters": [
          {
            "name": "incident_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncidentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Incident Status (To Selesai/Dialihkan)",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{incident_id}/intervene": {
      "post": {
        "operationId": "IncidentsController_interveneIncident",
        "parameters": [
          {
            "name": "incident_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterveneIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Intervene / Reassign Incident (Koordinator)",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/incidents/{incident_id}/rating": {
      "post": {
        "operationId": "IncidentsController_createIncidentRating",
        "parameters": [
          {
            "name": "incident_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Give Rating & Review to Finished Incident",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/my/incidents": {
      "get": {
        "operationId": "MyIncidentsController_getMyIncidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Track My Incident Reports (Warga Only)",
        "tags": [
          "My Incidents"
        ]
      }
    },
...
  },
  "info": {
    "title": "MANALINMAS API",
    "description": "Manajemen Kinerja Linmas & Pelaporan Darurat API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterWargaDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "phone_number": {
            "type": "string",
            "example": "08123456789"
          },
          "email": {
            "type": "string",
            "example": "budi@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "address": {
            "type": "string",
            "example": "Blok C3 No 12 RW 07"
          }
        },
        "required": [
          "full_name",
          "phone_number",
          "email",
          "password",
          "address"
        ]
      },
      "RegisterLinmasDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Andi Wijaya"
          },
          "phone_number": {
            "type": "string",
            "example": "08123456789"
          },
          "email": {
            "type": "string",
            "example": "andi@linmas.id"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "address": {
            "type": "string",
            "example": "Blok A2"
          },
          "employment_date": {
            "type": "string",
            "example": "2026-05-20"
          },
          "regu_id": {
            "type": "string",
            "example": "uuid-of-regu"
          }
        },
        "required": [
          "full_name",
          "phone_number",
          "email",
          "password",
          "address",
          "employment_date",
          "regu_id"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "andi@linmas.id"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "fcm_token": {
            "type": "string",
            "example": "firebase_device_token"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "jwt_refresh_token"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "fcm_token": {
            "type": "string",
            "example": "firebase_device_token"
          }
        },
        "required": [
          "fcm_token"
        ]
      },
...
      "ClockInDto": {
        "type": "object",
        "properties": {
          "shift_assignment_id": {
            "type": "string",
            "example": "uuid-of-shift-assignment"
          },
          "latitude": {
            "type": "number",
            "example": -6.90481
          },
          "longitude": {
            "type": "number",
            "example": 107.610313
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "Foto swafoto"
          }
        },
        "required": [
          "shift_assignment_id",
          "latitude",
          "longitude",
          "photo"
        ]
      },
      "ClockOutDto": {
        "type": "object",
        "properties": {
          "attendance_session_id": {
            "type": "string",
            "example": "uuid-of-attendance-session"
          },
          "latitude": {
            "type": "number",
            "example": -6.90481
          },
          "longitude": {
            "type": "number",
            "example": 107.610313
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "Foto swafoto"
          }
        },
        "required": [
          "attendance_session_id",
          "latitude",
          "longitude",
          "photo"
        ]
      },
...
      "CreateIncidentDto": {
        "type": "object",
        "properties": {
          "incident_type": {
            "type": "string",
            "example": "KEBAKARAN"
          },
          "title": {
            "type": "string",
            "example": "Rumah Pak Budi"
          },
          "description": {
            "type": "string",
            "example": "Api mulai membesar di dapur"
          },
          "latitude": {
            "type": "number",
            "example": -6.9683
          },
          "longitude": {
            "type": "number",
            "example": 107.5925
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "Foto Bukti"
          }
        },
        "required": [
          "incident_type",
          "title",
          "description",
          "latitude",
          "longitude",
          "photo"
        ]
      },
      "UpdateIncidentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "MENUNGGU",
              "DITANGANI",
              "SELESAI",
              "DIALIHKAN"
            ],
            "example": "SELESAI"
          },
          "resolution_notes": {
            "type": "string",
            "example": "Pelaku melarikan diri"
          }
        },
        "required": [
          "status",
          "resolution_notes"
        ]
      },
      "InterveneIncidentDto": {
        "type": "object",
        "properties": {
          "new_handler_linmas_id": {
            "type": "string",
            "example": "uuid-of-new-handler"
          },
          "intervention_reason": {
            "type": "string",
            "example": "Handler tidak merespons"
          }
        },
        "required": [
          "new_handler_linmas_id",
          "intervention_reason"
        ]
      },
      "CreateIncidentRatingDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "example": 5,
            "description": "Bintang 1-5"
          },
          "review": {
            "type": "string",
            "example": "Petugas datang dengan sangat cepat."
          }
        },
        "required": [
          "rating"
        ]
      },
...
    }
  }
}